backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
shopifyapi
cryptography
python-dotenv